+++ src/main/java/org/elasticsearch/action/search/type/TransportSearchScrollQueryAndFetchAction.java
@@ -149,7 +149,6 @@ public class TransportSearchScrollQueryAndFetchAction extends AbstractComponent 
                     if (counter.decrementAndGet() == 0) {
                         finishHim();
                     }
-                } else {
                 }
             }
         }
